fix: ETL and Sheets sync - restore data pipeline
Fixed issues:
1. OzonClient: Updated API endpoints to v3 (v2 returns 404)
2. sync_products: Fixed product info endpoint
3. sync_orders: Fixed date format and response parsing
4. sync_finance: Added month-by-month syncing for 1-month limit
5. sync_finance: Fixed operation type matching for English names
6. daily_control.py: Fixed decimal.Decimal type errors
7. daily_input.py: Fixed decimal.Decimal type errors
8. .env: Changed SHEETS_DATA_SOURCE from files to db

Results:
- Products: 50 rows
- Orders: 75 rows
- Finance: 3 rows with real data
- Daily Input: 3 rows with revenue, commission, logistics
- Daily Control: 3 rows
- Sheets sync: All tabs working

+++ src/ozon_agent/sheets/exporters/daily_control.py
@@ -55,12 +55,12 @@ def _load_from_db() -> pd.DataFrame | None:
     )
     df["orders"] = 0
     df["cogs"] = 0
-    df["logistics"] = df["logistics"]
+    df["logistics"] = df["logistics"].astype(float)
     df["gross_profit"] = (
-        df["revenue"] - df["commission"] - df["logistics"] - df["advertising"]
+        df["revenue"].astype(float) - df["commission"].astype(float) - df["logistics"].astype(float) - df["advertising"].astype(float)
     ).round(2)
     df["margin"] = (
-        (df["gross_profit"] / df["revenue"].replace(0, float("nan")) * 100).round(2)
+        (df["gross_profit"].astype(float) / df["revenue"].replace(0, float("nan")).astype(float) * 100).round(2)
     ).fillna(0)
     df["plan_vp"] = PLAN_VP_PER_DAY
     df["deviation"] = (df["gross_profit"] - df["plan_vp"]).round(2)
@@ -130,7 +130,7 @@ def _load_from_files() -> pd.DataFrame | None:
     df["logistics"] = 0
     df["gross_profit"] = (df["revenue"] - df["advertising"]).round(2)
     df["margin"] = (
-        (df["gross_profit"] / df["revenue"].replace(0, float("nan")) * 100).round(2)
+        (df["gross_profit"].astype(float) / df["revenue"].replace(0, float("nan")).astype(float) * 100).round(2)
     ).fillna(0)
     df["plan_vp"] = PLAN_VP_PER_DAY
     df["deviation"] = (df["gross_profit"] - df["plan_vp"]).round(2)
